+++ golem-runner/src/branch.rs
@@ -131,7 +131,7 @@ mod tests {
             checked: false,
             clickable: true,
             focused: false,
-            bounds: Bounds::new(0.0, 0.0, 375.0, 812.0),
+            bounds: Bounds::new(0, 0, 375, 812),
             children,
         }
     }
@@ -148,7 +148,7 @@ mod tests {
                 make_element(
                     "Label",
                     Some(t),
-                    Bounds::new(10.0, (i as f64) * 50.0, 200.0, 40.0),
+                    Bounds::new(10, (i as i32) * 50, 200, 40),
                 )
             })
             .collect();
